+++ app/Main.hs
@@ -11,12 +11,11 @@ import Text.Pretty.Simple (pPrint)
 import qualified Lang as L0
 
 [deflang| L0.L0 :-> L1
-  (*
-    (Expr
-      (- Lam)
-      (+ Lam {x String} {e $Expr})
-      (- Nope)
-    ))
+  (* Expr
+    (- Lam)
+    (+ Lam {x String} {e $Expr})
+    (- Nope)
+  )
   (- Stmt)
 |]
 deriving stock instance Show Expr